Reissner-Nordstorm metric
Added Reissner-Nordstorm metric to vaccum_solutions.py
