+++ src/einsteinpy/symbolic/predefined/vacuum_solutions.py
@@ -15,9 +15,9 @@ def Schwarzschild(c=constants.c, sch=symbols("r_s")):
     c : ~sympy.core.basic.Basic or int or float
         Any value to assign to speed of light. Defaults to ``c``.
     sch : ~sympy.core.basic.Basic or int or float
-        Any value to assign to Schwarzschild Radius of the central object. 
+        Any value to assign to Schwarzschild Radius of the central object.
         Defaults to ``r_s``.
-    
+
     """
     coords = symbols("t r theta phi")
     t, r, theta, phi = coords
@@ -37,7 +37,7 @@ def Kerr(c=constants.c, sch=symbols("r_s"), a=symbols("a")):
     c : ~sympy.core.basic.Basic or int or float
         Any value to assign to speed of light. Defaults to ``c``.
     sch : ~sympy.core.basic.Basic or int or float
-        Any value to assign to Schwarzschild Radius of the central object. 
+        Any value to assign to Schwarzschild Radius of the central object.
         Defaults to ``r_s``.
     a : ~sympy.core.basic.Basic or int or float
         Spin factor of the heavy body. Usually, given by ``J/(Mc)``,
@@ -85,14 +85,14 @@ def KerrNewman(
     eps_0 : ~sympy.core.basic.Basic or int or float
         Any value to assign to the electric constant or permittivity of free space. Defaults to ``eps_0``.
     sch : ~sympy.core.basic.Basic or int or float
-        Any value to assign to Schwarzschild Radius of the central object. 
+        Any value to assign to Schwarzschild Radius of the central object.
         Defaults to ``r_s``.
     a : ~sympy.core.basic.Basic or int or float
         Spin factor of the heavy body. Usually, given by ``J/(Mc)``,
         where ``J`` is the angular momentum.
         Defaults to ``a``.
     Q:  ~sympy.core.basic.Basic or int or float
-        Any value to assign to eletric charge of the central object. 
+        Any value to assign to eletric charge of the central object.
         Defaults to ``Q``.
 
     """
@@ -114,4 +114,49 @@ def KerrNewman(
         / c2,
     ).tolist()
     metric[0][3] = metric[3][0] = sch * r * a * (sin(theta) ** 2) / (Sigma * c)
+    return MetricTensor(metric, coords, "ll")
+
+def Reissner_Nordstorm(
+    c=constants.c,
+    G=constants.G,
+    eps_0=constants.eps_0,
+    sch=symbols("r_s"),
+    a=symbols("a"),
+    Q=symbols("Q"),
+):
+    """"
+    The Reissner–Nordström metric in spherical coordinates
+
+    A static solution to the Einstein–Maxwell field equations,
+    which corresponds to the gravitational field of a charged,
+    non-rotating, spherically symmetric body of mass M.
+
+    Parameters
+    ----------
+    c : ~sympy.core.basic.Basic or int or float
+        Any value to assign to speed of light. Defaults to ``c``.
+    G : ~sympy.core.basic.Basic or int or float
+        Any value to assign to the Newton's (or gravitational) constant. Defaults to ``G``.
+    eps_0 : ~sympy.core.basic.Basic or int or float
+        Any value to assign to the electric constant or permittivity of free space. Defaults to ``eps_0``.
+    sch : ~sympy.core.basic.Basic or int or float
+        Any value to assign to Schwarzschild Radius of the central object.
+        Defaults to ``r_s``.
+    a : ~sympy.core.basic.Basic or int or float
+        Spin factor of the heavy body. Usually, given by ``J/(Mc)``,
+        where ``J`` is the angular momentum.
+        Defaults to ``a``.
+    Q:  ~sympy.core.basic.Basic or int or float
+        Any value to assign to eletric charge of the central object.
+        Defaults to ``Q``.
+
+    """
+    coords = symbols("t r theta phi")
+    t, r, theta, phi = coords
+    rQsq = ((Q ** 2) * G) / (4 * pi * eps_0 * (c ** 4))
+    Arn = 1 - sch / r + rQsq / r ** 2
+
+    metric = diag(
+        (Arn * (c ** 2)), -(1 / Arn), -(r ** 2), -(r ** 2) * sin(theta) ** 2,
+    ).tolist()
     return MetricTensor(metric, coords, "ll")
